use reshape

+++ ppgan/models/lapstyle_model.py
@@ -822,7 +822,7 @@ class LapStyleRevFirstThumb(BaseModel):
         self.loss_Dp_real = 0
         spshape = self.sp.shape
         print(spshape)
-        reshaped = self.sp.reshape((4,spshape[0],spshape[1],spshape[2]/2,spshape[3]/2))
+        reshaped = self.sp.reshape((4,int(spshape[0]),int(spshape[1]),int(spshape[2]/2),int(spshape[3]/2)))
         for i in range(4):
             pred_p_real = self.nets['netD_patch'](reshaped[i])
             self.loss_Dp_real += self.gan_criterion(pred_p_real, True)
